Update dpvar subsagn.
The function now supports adjusting elements of 'dpvar' objects using linear indexing, so that e.g. `P(i) = b` for a mxn 'dpvar' object `P` sets the ith element of this object equal to `b`, where `b` can be of type 'double', 'polynomial', or 'dpvar'. If `P` is vector-valued, so that either m=1 or n=1, `b` can also be set to `[]` to remove the ith element from `P`. The index `i` can also be vector-valued, so long as the number of elements of `b` matches that of `i`.