Update sossolve.m
A bug has been fixed in the case the solver SDPT3 is used. In converting the program to SDPT3 format, an option `smallblkdim=60` was used (for processing the positivity constraints), but in converting the solved program back to SeDuMi format, this was replaced with `smblkdim=100`. This discrepancy caused an error for systems involving positive matrices of size larger than 60x60 but smaller than 100x100. This has been resolved, now using `smallblkdim=smblkdim=60`.