Update dpvar.m
In converting a cell of decision variable names to an array of decision variables ('dpvar' object), the function previously did not account for the fact that the same decision variable name may be used in different elements of the cell. As such, the field 'dvarname' of the output 'dpvar' object could contain duplicate variable names, potentially causing problems later on. This has been resolved, now only storing unique variable names in the returned 'dpvar' object.