Update sosgetsol.m
When evaluating the solved value of a `dpvar` object, the function previously did not take into account that the list of coefficients appearing in the object may not be unique, properly setting the value of one instance of a coefficient, but not of all other instances of this same coefficients. This has been resolved, combining the coefficient names into a unique list, so that all coefficients are now properly substituted.